policy: close voip output when new output is opened by same voip client
When USB headset is inserted, VoIP client will stop current output
and then open a new output. Due to voip usecase is not closed in time,
the new output is opened in handset with FLAG_NONE and fallback to
primary output.

Fix it by closing voip output when a new output is opened by same voip
client in different device.

CRs-Fixed: 2272627
Change-Id: I71672675f13f58ed818117d5d0abe0d93204227e